refactor: 사용하는 위치에 따라 dto 패키지 분리
