+++ src/main/java/chess/dao/GameDao.java
@@ -1,7 +1,7 @@
 package chess.dao;
 
-import chess.dto.GameDto;
-import chess.dto.GameStatusDto;
+import chess.service.dto.GameDto;
+import chess.service.dto.GameStatusDto;
 
 public interface GameDao {
