+++ src/main/java/chess/service/dto/GameStatusDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.service.dto;
 
 public enum GameStatusDto {
