+++ src/main/java/chess/service/dto/PieceDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.service.dto;
 
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceColor;
