+++ src/test/java/chess/dao/FakeGameDao.java
@@ -1,8 +1,8 @@
 package chess.dao;
 
 
-import chess.dto.GameDto;
-import chess.dto.GameStatusDto;
+import chess.service.dto.GameDto;
+import chess.service.dto.GameStatusDto;
 
 public class FakeGameDao implements GameDao {
