+++ src/test/java/chess/service/ChessServiceTest.java
@@ -4,9 +4,8 @@ import chess.dao.FakeGameDao;
 import chess.dao.FakePieceDao;
 import chess.domain.command.MoveCommand;
 import chess.domain.piece.PieceFactory;
-import chess.dto.ChessResponseDto;
-import chess.dto.PieceDto;
-import chess.serviece.ChessService;
+import chess.service.dto.ChessResponseDto;
+import chess.service.dto.PieceDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
